+++ packages/components/tabs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @webbo-ui/tabs
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @webbo-ui/theme@1.0.3
+
 ## 1.0.3
 
 ### Patch Changes